Moved Semantic Maps class into dedicated file
Probably adresses https://github.com/JeroenDeDauw/Maps/issues/286
